backend/requirements.txt
Django==4.2.10
asgiref==3.7.2
sqlparse==0.4.4
django-cors-headers==4.3.1
djangorestframework==3.14.0